Adds digital pack rollout for order summary and tweaks suggested by Rupert
