+++ support-frontend/assets/components/subscriptionCheckouts/summary.module.scss
@@ -4,19 +4,45 @@
   position: relative;
 }
 
-.content {
-  padding: $gu-v-spacing/2 $gu-h-spacing/2 $gu-v-spacing;
+.content, .contentShowDetails {
   @include mq($until: tablet) {
-    background: gu-colour(neutral-97);
     display: grid;
-    grid-template-columns: max-content minmax(50%, min-content);
+    grid-template-columns: 40% 60%;
     justify-content: space-between;
+    text-align: right;
   }
   @include mq($from: tablet) {
     & > * + * {
       margin-top: $gu-v-spacing;
     }
   }
+  @include mq($until: mobileMedium) {
+    grid-template-columns: 50% 50%;
+  }
+}
+
+.contentWrapper, .content {
+  padding: $gu-v-spacing/2 $gu-h-spacing/2 $gu-v-spacing;
+  @include mq($until: tablet) {
+    background: gu-colour(neutral-97);
+  }
+}
+
+.contentShowDetails, .contentShowDetailsLast {
+  grid-template-columns: 40% 55%;
+  text-align: left;
+  border-bottom: 1px solid gu-colour(neutral-86);
+  margin-bottom: $gu-v-spacing *2 /3;
+  padding-bottom: $gu-v-spacing;
+}
+
+.contentShowDetailsLast {
+  border-bottom: none;
+  display: grid;
+  grid-template-columns: 40% 55%;
+  padding-bottom: 0;
+  margin-top: $gu-v-spacing;
+  margin-bottom: 0;
 }
 
 .img {
@@ -35,7 +61,7 @@
   display: block;
 }
 
-.header {
+.header, .headerShowDetails {
   @include gu-fontset-body-sans;
   font-weight: 600;
   @include mq($from: tablet) {
@@ -50,9 +76,17 @@
     position: absolute;
     top: $gu-v-spacing;
   }
+  @include mq($until: tablet) {
+    margin-bottom: $gu-v-spacing /2;
+    text-align: left;
+  }
 }
 
-.title {
+.headerShowDetails {
+  margin-bottom: $gu-v-spacing;
+}
+
+.title, .titleLeftAlign {
   @include gu-fontset-heading;
   @include mq($until: tablet) {
     @include gu-typeface(headline, 18, $size-only: true);
@@ -61,18 +95,54 @@
     @include gu-typeface(headline, 32, $size-only: true);
   }
 }
+
 .titleDescription {
   @include gu-fontset-body-sans;
+}
+
+.titleLeftAlign {
+  text-align: left;
+}
+
+.dropDown, .changeSub {
+  @include gu-fontset-body-sans;
+  outline: none;
+  border: none;
+  padding: 0;
+  text-align: left;
+}
+
+.changeSub, :visited {
+  color: gu-colour(neutral-7);
+  text-align: right;
+  right: 0;
+}
+
+.pricing {
+  @include gu-fontset-body-sans;
+  font-size: 14px;
+  font-weight: 600;
+  @include mq($from: tablet) {
+    @include gu-fontset-heading;
+    font-weight: 400;
+  }
+}
+
+.dataList {
+  @include mq($from: tablet) {
+    margin-bottom: $gu-v-spacing;
+  }
+}
+
+.tabletAndDesktop {
   @include mq($until: tablet) {
     display: none;
   }
 }
 
-.pricing {
-  @include gu-fontset-body-sans;
+.mobileOnly {
   @include mq($from: tablet) {
-    @include gu-fontset-heading;
-    font-weight: 400;
+    display: none;
   }
 }
 
@@ -80,14 +150,29 @@
   @include gu-fontset-body-sans;
   display: grid;
   grid-template-columns: max-content 1fr;
-  @include mq($until: tablet) {
-    display: none;
-  }
 }
 
-.data dt {
+.data dt, .dataUnderline, .dataBold {
   font-weight: bold;
   padding-right: $gu-h-spacing;
+}
+
+.dataBold {
+  font-weight: 600;
+}
+
+.spaceRight {
+  padding-right: $gu-h-spacing /3;
+}
+
+.defaultState {
+  transition: transform 300ms;
+  transform-origin: center;
+  transform: rotate(180deg);
+}
+
+.openState {
+  transform: rotate(0);
 }
 
 .promo {
